refactor - folder restructuring, filename modification, authenticationSlice state change
