+++ src/redux/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import authReducer from "../slices/authenticationSlice";
+import authReducer from "./slices/authenticationSlice";
 
 export const store = configureStore({
   reducer: {
